feat: :recycle: make flattened schemas and types for them reusable thru internal packages
