+++ apps/backend/src/controllers/todayHighlightController.ts
@@ -3,11 +3,11 @@ import axios from 'axios';
 import { prisma } from '@/prismaClient';
 import { compareTime, flattenTodayData } from '../helpers/controllerHelpers';
 import {
-  TodayHighlightSchema,
-  TodayHiglightSchemaType,
-  FlattenedTodayHighlightSchemaType,
-  FlattenedTodayHighlightSchema,
-} from '../../schema/todayHighlightSchema';
+  UnFlattenedTodayHighlightSchemaType,
+  UnFlattenedTodayHighlightSchema,
+} from '@/schema/weatherApi';
+import { TodayHighlightSchema } from 'shared-schemas/apiSchemas';
+import { TodayHighlightSchemaType } from 'shared-types/apiTypes';
 
 export const getTodayHighlight = async (
   req: Request,
@@ -34,13 +34,13 @@ export const getTodayHighlight = async (
     },
   });
 
-  const parsedData = TodayHighlightSchema.parse(data);
+  const parsedData = UnFlattenedTodayHighlightSchema.parse(data);
   const flattenedData = flattenTodayData<
-  TodayHiglightSchemaType,
-  FlattenedTodayHighlightSchemaType
+    UnFlattenedTodayHighlightSchemaType,
+    TodayHighlightSchemaType
   >(parsedData);
 
-  const validatedData = FlattenedTodayHighlightSchema.parse(flattenedData);
+  const validatedData = TodayHighlightSchema.parse(flattenedData);
 
   if (highlightData) {
     await prisma.highlightData.create({
