+++ packages/shared-schemas/apiSchemas.ts
@@ -9,7 +9,7 @@ export const FetchTodayOverviewResult = z.object({
   location: z.object({
     name: z.string(),
     country: z.string(),
-    localtime: z.string(),
+    localtime: z.string().transform((val) => new Date(val).toISOString()),
   }),
   current: z.object({
     temp_c: z.number(),
@@ -42,4 +42,24 @@ export const FetchForecastResult = z.object({
       })
     ),
   }),
+});
+
+export const TodayHighlightSchema = z.object({
+  localtime: z.string().transform((val) => new Date(val).toISOString()),
+  humidity: z.number(),
+  uv: z.number(),
+  vis_km: z.number(),
+  wind_kph: z.number(),
+  precip_mm: z.number(),
+  heatindex_c: z.number(),
+});
+
+export const CurrentWeatherSchema = z.object({
+  name: z.string(),
+  country: z.string(),
+  localtime: z.string().transform((val) => new Date(val).toISOString()),
+  temp_c: z.number(),
+
+  condition: z.string(),
+  icon: z.string(),
 });
