+++ packages/shared-types/apiTypes.ts
@@ -2,6 +2,8 @@ import { z } from 'zod';
 import {
   FetchForecastResult,
   FetchTodayOverviewResult,
+  TodayHighlightSchema,
+  CurrentWeatherSchema,
 } from 'shared-schemas/apiSchemas';
 
 export type CityName = z.infer<
@@ -10,3 +12,7 @@ export type CityName = z.infer<
 export type WeatherData = z.infer<typeof FetchTodayOverviewResult>;
 
 export type ForecastData = z.infer<typeof FetchForecastResult>;
+
+export type TodayHighlightSchemaType = z.infer<typeof TodayHighlightSchema>;
+
+export type CurrentWeatherSchemaType = z.infer<typeof CurrentWeatherSchema>;
